requirements.txt
Cython==0.29.24
numpy==1.19.5
joblib==1.0.1
pandas==1.1.5
Pillow==8.3.1
scikit-learn==0.24.2
streamlit==0.85.0
psycopg2-binary==2.8.6
plotly==5.4.0
